ajout petit

+++ Project_V3_SIAM_PASCAL_GERONDEAU/include/Mountain.hpp
@@ -23,7 +23,7 @@ class Mountain : virtual public Piece
         //METHODS
         std::string Getstring();
         void display(BITMAP* dest);
-        int push(BoardGame& board,char direction,char order, int power_sum, bool first);
+        int push(BoardGame& board,char direction,char order, float power_sum, bool first);
         void SetOrientation(int x){} //Unused virtual function
