+++ Project_V3_SIAM_PASCAL_GERONDEAU/include/Pawn.hpp
@@ -30,7 +30,7 @@ class Pawn :  public Piece
         //METHODS
 
         void display(BITMAP* dest);
-        int push(BoardGame& board,char direction,char order, int power_sum, bool first);
+        int push(BoardGame& board,char direction,char order, float power_sum, bool first);
         std::string Getstring();
